Create dgim.utils
Move generate_random_stream in it.
